+++ src/com/kaylerrenslow/armaDialogCreator/control/ControlPropertyLookup.java
@@ -8,7 +8,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- A place to find ALL known control properties for all controls. This is where the name of the property, property type, description, and options (if allowed) are listed.
+ A place to find ALL known control properties for all controls.
+ This is where the name of the property, property type, description, and options (if allowed) are listed.
 
  @author Kayler
  @since 05/22/2016. */
@@ -41,141 +42,46 @@ public enum ControlPropertyLookup implements ControlPropertyLookupConstant {
 	BLINKING_PERIOD(23, "blinkingPeriod", PropertyType.Float),
 
 	/*Static*/
-	STATIC_AUTO_PLAY(24, "autoPlay", PropertyType.Boolean),
-	STATIC_KEY(25, "key", PropertyType.String),
-	STATIC_LOOPS(26, "loops", PropertyType.Int),
-	STATIC_LINE_SPACING(27, "lineSpacing", PropertyType.Float),
-	STATIC_FIXED_WIDTH(28, "fixedWidth", PropertyType.Boolean),
+	AUTO_PLAY(24, "autoPlay", PropertyType.Boolean),
+	KEY(25, "key", PropertyType.String),
+	LOOPS(26, "loops", PropertyType.Int),
+	LINE_SPACING(27, "lineSpacing", PropertyType.Float),
+	FIXED_WIDTH(28, "fixedWidth", PropertyType.Boolean),
 
 	/*Structured Text*/
-	STRUCT_TEXT_SIZE(29, "size", PropertyType.Float),
+	SIZE(29, "size", PropertyType.Float),
 
 	/*HTML*/
-	HTML_CYCLE_LINKS(30, "cyclelinks", PropertyType.Boolean),
-	HTML_FILE_NAME(31, "filename", PropertyType.FileName),
-	HTML_COLOR_BOLD(32, "colorBold", PropertyType.Color),
-	HTML_COLOR_LINK(33, "colorLink", PropertyType.Color),
-	HTML_COLOR_LINK_ACTIVE(34, "colorLinkActive", PropertyType.Color),
-	HTML_COLOR_PICTURE(35, "colorPicture", PropertyType.Color),
-	HTML_COLOR_PICTURE_BORDER(36, "colorPictureBorder", PropertyType.Color),
-	HTML_COLOR_PICTURE_LINK(37, "colorPictureLink", PropertyType.Color),
-	HTML_COLOR_PICTURE_SELECTED(38, "colorPictureSelected", PropertyType.Color),
-	HTML_PREV_PAGE(39, "prevPage", PropertyType.Image),
-	HTML_NEXT_PAGE(40, "nextPage", PropertyType.Image),
+	CYCLE_LINKS(30, "cycleLinks", PropertyType.Boolean),
+	FILE_NAME(31, "filename", PropertyType.FileName),
+	COLOR_BOLD(32, "colorBold", PropertyType.Color),
+	COLOR_LINK(33, "colorLink", PropertyType.Color),
+	COLOR_LINK_ACTIVE(34, "colorLinkActive", PropertyType.Color),
+	COLOR_PICTURE(35, "colorPicture", PropertyType.Color),
+	COLOR_PICTURE_BORDER(36, "colorPictureBorder", PropertyType.Color),
+	COLOR_PICTURE_LINK(37, "colorPictureLink", PropertyType.Color),
+	COLOR_PICTURE_SELECTED(38, "colorPictureSelected", PropertyType.Color),
+	PREV_PAGE(39, "prevPage", PropertyType.Image),
+	NEXT_PAGE(40, "nextPage", PropertyType.Image),
 
 	/*Button*/
-	BTN_ACTION(41, "action", PropertyType.SQF),
-	BTN_BORDER_SIZE(42, "borderSize", PropertyType.Float),
-	BTN_COLOR_BACKGROUND_ACTIVE(43, "colorBackgroundActive", PropertyType.Color),
-	BTN_COLOR_BACKGROUND_DISABLED(44, "colorBackgroundDisabled", PropertyType.Color),
-	BTN_COLOR_BORDER(45, "colorBorder", PropertyType.Color),
-	BTN_COLOR_DISABLED(46, "colorDisabled", PropertyType.Color),
-	BTN_COLOR_FOCUSED(47, "colorFocused", PropertyType.Color),
-	BTN_COLOR_SHADOW(48, "colorShadow", PropertyType.Color),
-	BTN_DEFAULT(49, "default", PropertyType.Boolean),
-	BTN_OFFSET_PRESSED_X(50, "offsetPressedX", PropertyType.Float),
-	BTN_OFFSET_PRESSED_Y(51, "offsetPressedY", PropertyType.Float),
-	BTN_OFFSET_X(52, "offsetX", PropertyType.Float),
-	BTN_OFFSET_Y(53, "offsetY", PropertyType.Float),
+	ACTION(41, "action", PropertyType.SQF),
+	BORDER_SIZE(42, "borderSize", PropertyType.Float),
+	COLOR_BACKGROUND_ACTIVE(43, "colorBackgroundActive", PropertyType.Color),
+	COLOR_BACKGROUND_DISABLED(44, "colorBackgroundDisabled", PropertyType.Color),
+	COLOR_BORDER(45, "colorBorder", PropertyType.Color),
+	COLOR_DISABLED(46, "colorDisabled", PropertyType.Color),
+	COLOR_FOCUSED(47, "colorFocused", PropertyType.Color),
+	COLOR_SHADOW(48, "colorShadow", PropertyType.Color),
+	DEFAULT(49, "default", PropertyType.Boolean),
+	OFFSET_PRESSED_X(50, "offsetPressedX", PropertyType.Float),
+	OFFSET_PRESSED_Y(51, "offsetPressedY", PropertyType.Float),
+	OFFSET_X(52, "offsetX", PropertyType.Float),
+	OFFSET_Y(53, "offsetY", PropertyType.Float),
 	SOUND_CLICK(54, "soundClick", PropertyType.Sound),
 	SOUND_ENTER(55, "soundEnter", PropertyType.Sound),
 	SOUND_ESCAPE(56, "soundEscape", PropertyType.Sound),
 	SOUND_PUSH(57, "soundPush", PropertyType.Sound),
-
-	/* ..Shortcut Button */
-	BTN_ANIM_TEXTURE_NORMAL(58, "animTextureNormal", PropertyType.Texture),
-	BTN_ANIM_TEXTURE_DISABLED(59, "animTextureDisabled", PropertyType.Texture),
-	BTN_ANIM_TEXTURE_OVER(60, "animTextureOver", PropertyType.Texture),
-	BTN_ANIM_TEXTURE_FOCUSED(61, "animTextureFocused", PropertyType.Texture),
-	BTN_ANIM_TEXTURE_PRESSED(62, "animTexturePressed", PropertyType.Texture),
-	BTN_ANIM_TEXTURE_DEFAULT(63, "animTextureDefault", PropertyType.Texture),
-	BTN_TEXTURE_NO_SHORTCUT(64, "textureNoShortcut", PropertyType.Texture),
-	BTN_COLOR2(65, "color2", PropertyType.Color),
-	BTN_COLOR_BACKGROUND2(66, "colorBackground2", PropertyType.Color),
-	BTN_PERIOD(67, "period", PropertyType.Float),
-	BTN_PERIOD_FOCUS(68, "periodFocus", PropertyType.Float),
-	BTN_PERIOD_OVER(69, "periodOver", PropertyType.Float),
-
-	/*Active Text*/
-	AT_ACTION(70, "action", PropertyType.SQF),
-	AT_CAN_DRAG(71, "canDrag", PropertyType.Boolean),
-	AT_COLOR(72, "color", PropertyType.Color),
-	AT_COLOR_ACTIVE(73, "colorActive", PropertyType.Color),
-	AT_COLOR_SHADE(74, "colorShade", PropertyType.Color),
-	AT_COLOR_FOCUSED(75, "colorFocused", PropertyType.Color),
-	AT_COLOR_DISABLED(76, "colorDisabled", PropertyType.Color),
-	AT_COLOR_BACKGROUND2(77, "colorBackground2", PropertyType.Color),
-	AT_DEFAULT(78, "default", PropertyType.Boolean),
-	AT_PICTURE_WIDTH(79, "pictureWidth", PropertyType.Float),
-	AT_PICTURE_HEIGHT(80, "pictureHeight", PropertyType.Float),
-	AT_SIDE_DISABLED(81, "sideDisabled", PropertyType.Color),
-	AT_PICTURE(82, "picture", PropertyType.Texture),
-	AT_SIDE_TOGGLE(83, "sideToggle", PropertyType.Color),
-	AT_TEXT_HEIGHT(84, "textHeight", PropertyType.Float),
-
-	/*Edit*/
-	EDIT_AUTO_COMPLETE(85, "autocomplete", PropertyType.String, new ControlPropertyOption("None", "", "No auto-completion."), new ControlPropertyOption("Script", "scripting", "Auto-completion set for scripting."), new ControlPropertyOption("General", "general", "Auto-completion on normal words.")),
-	EDIT_HTML_CONTROL(86, "htmlControl", PropertyType.Boolean),
-	EDIT_LINE_SPACING(87, "lineSpacing", PropertyType.Float),
-	EDIT_COLOR_SELECTION(88, "colorSelection", PropertyType.Color),
-	EDIT_SIZE(89, "size", PropertyType.Float),
-
-	/*Sliders*/
-	SLIDE_ARROW_EMPTY(90, "arrowEmpty", PropertyType.Texture),
-	SLIDE_ARROW_FULL(91, "arrowFull", PropertyType.Texture),
-	SLIDE_BORDER(92, "border", PropertyType.Texture),
-	SLIDE_ACTIVE(93, "colorActive", PropertyType.Color),
-	SLIDE_DISABLED(94, "colorDisabled", PropertyType.Color),
-	SLIDE_THUMB(95, "thumb", PropertyType.Texture),
-	SLIDE_VSPACING(96, "vspacing", PropertyType.Float),
-
-	/*Combo*/
-	COMBO_ARROW_EMPTY(97, "arrowEmpty", PropertyType.Texture),
-	COMBO_ARROW_FULL(98, "arrowFull", PropertyType.Texture),
-	COMBO_COLOR(99, "color", PropertyType.Color),
-	COMBO_ACTIVE(100, "colorActive", PropertyType.Color),
-	COMBO_DISABLED(101, "colorDisabled", PropertyType.Color),
-	COMBO_COLOR_SCROLL_BAR(102, "colorScrollbar", PropertyType.Color),
-	COMBO_COLOR_SELECT(103, "colorSelect", PropertyType.Color),
-	COMBO_COLOR_SELECT_BACKGROUND(104, "colorSelectBackground", PropertyType.Color),
-	COMBO_FROM(105, "from", PropertyType.Float),
-	COMBO_TO(106, "to", PropertyType.Float),
-	COMBO_MAX_HISTORY_DELAY(107, "maxHistoryDelay", PropertyType.Float),
-	COMBO_ROW_HEIGHT(108, "rowHeight", PropertyType.Float),
-	COMBO_SOUND_SELECT(109, "soundSelect", PropertyType.Sound),
-	COMBO_SOUND_COLLAPSE(110, "soundCollapse", PropertyType.Sound),
-
-	/*Listboxes*/
-	LB_ACTIVE(111, "active", PropertyType.Boolean),
-	LB_AUTO_SCROLL(112, "autoScroll", PropertyType.Int),
-	LB_ARROW_EMPTY(113, "arrowEmpty", PropertyType.Texture),
-	LB_ARROW_FULL(114, "arrowFull", PropertyType.Texture),
-	LB_BORDER(115, "border", PropertyType.Texture),
-	LB_CAN_DRAG(116, "canDrag", PropertyType.Boolean),
-	LB_COLLISION_COLOR(117, "collisionColor", PropertyType.Color),
-	LB_COLOR_SCROLL_BAR(118, "colorScrollbar", PropertyType.Color),
-	LB_COLOR(119, "color", PropertyType.Color),
-	LB_COLOR_PLAYER_ITEM(120, "colorPlayerItem", PropertyType.Color),
-	LB_COLOR_ACTIVE(121, "colorActive", PropertyType.Color),
-	LB_COLOR_DISABLED(122, "colorDisabled", PropertyType.Color),
-	LB_COLOR_SELECT(123, "colorSelect", PropertyType.Color),
-	LB_COLOR_SELECT2(124, "colorSelect2", PropertyType.Color),
-	LB_COLOR_SELECT_BACKGROUND(125, "colorSelectBackground", PropertyType.Color),
-	LB_COLOR_SELECT_BACKGROUND2(126, "colorSelectBackground2", PropertyType.Color),
-	LB_COLUMNS(127, "columns", PropertyType.Array),
-	LB_DISABLED(128, "disabled", PropertyType.Boolean),
-	LB_ENABLED(129, "enabled", PropertyType.Boolean),
-	LB_DISABLED_CTRL_COLOR(130, "disabledCtrlColor", PropertyType.Color),
-	LB_DISABLED_KEY_COLOR(131, "disabledKeyColor", PropertyType.Color),
-	LB_DRAW_SIDE_ARROWS(132, "drawSideArrows", PropertyType.Boolean),
-	LB_IDC_LEFT(133, "idcLeft", PropertyType.Int),
-	LB_IDC_RIGHT(134, "idcRight", PropertyType.Int),
-	LB_MAIN_COLUMNW(135, "mainCollumW", PropertyType.Float),
-	LB_SECOND_COLUMNW(136, "secndCollumW", PropertyType.Float),
-	LB_LINE_SPACING(138, "lineSpacing", PropertyType.Float),
-	LB_MAX_HISTORY_DELAY(139, "maxHistoryDelay", PropertyType.Float),
-	LB_ROW_HEIGHT(140, "rowHeight", PropertyType.Float),
-	LB_ROWS(141, "rows", PropertyType.Int),
 
 	/*event handlers*/
 	EVENT_ON_LOAD(1000, "onLoad", PropertyType.SQF),
@@ -228,18 +134,7 @@ public enum ControlPropertyLookup implements ControlPropertyLookupConstant {
 	EVENT_ON_OBJECT_MOVED(1047, "onObjectMoved", PropertyType.SQF),
 	EVENT_ON_MENU_SELECTED(1048, "onMenuSelected", PropertyType.SQF),
 	EVENT_ON_DRAW(1049, "onDraw", PropertyType.SQF),
-	EVENT_ON_VIDEO_STOPPED(1050, "onVideoStopped", PropertyType.SQF),
-
-
-	/*HitZone and TextPos classes*/
-	/*TOP(0, "top", PropertyType.Float),
-	RIGHT(0, "right", PropertyType.Float),
-	BOTTOM(0, "bottom", PropertyType.Float),
-	LEFT(0, "left", PropertyType.Float),
-
-	/*ShortcutPos class*/
-	/*CLASS_SHORTCUT_POS__W(0, "w", PropertyType.Float),
-	CLASS_SHORTCUT_POS__H(0, "h", PropertyType.Float)*/;
+	EVENT_ON_VIDEO_STOPPED(1050, "onVideoStopped", PropertyType.SQF),;
 
 
 	public static final ReadOnlyList<ControlPropertyLookup> EMPTY = new ReadOnlyList<>(new ArrayList<>());