+++ CHANGES.md
@@ -1,11 +1,12 @@
 **Added:**
-* frame render support
+* event properties to static
+* event properties documentation
 
 **Changed:**
-* 
+* cleaned up ControlPropertyLookup
 
 **Fixed:**
-* 
+* error when fileName for getting image had a trailing space
 
 **Notes:**
 * May 31: we should have default value providers in one xml file for every control.